requirements.txt
streamlit
openai
azure-storage-blob
azure-communication-email
reportlab
fpdf2
plotly
pandas
Pillow
python-dotenv
requests
